requirements.txt
Flask==3.0.2
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
gdown==5.1.0
waitress==3.0.0